Make animation time based and convert units.
Metres to milimetres; frames to miliseconds.

With animation lengths not being restricted to an integral frame count, balls are no longer guaranteed to be forever synchronised (rounding errors accumulation). An easy fix is tracking the total time of animation (`fullPeriod * beatDuration`) and synchronise ball timers with that on each cycle. I didn't manage to produce the issue though.