Add GIF exporting.
By exposing:
- `Animator.toGIF(siteswap, settings)`
- `Animator.prototype.toGIF(settings)`

They return a promise that resolves with the GIF blob. The promise's `.then(onResolve, onReject, onProgress)` accepts a third callback parameter that gets called with progress percentage ([0-1] range) as frames are processed.

`settings` are forwarded to `Animator.prototype.configure()`, but also support some GIF specific configuration:
- width
- height
- fps
- backgroundColor

The operation is quite slow and resource intensive. The `gif.js` library kind of sucks but it's the best I could find:
- all frames are kept in memory before encoding starts (I should probably move this step into a worker)
- identical frames are not reused
- frames do not use a global palette
- no metadata support
- huge distribution size ~30kb, doubling the project size